Add Mushr Extension to Marketplace (#2)
**Public-Facing Changes**
Add a small extension to the marketplace with publishing buttons for the MuSHR package. See [this repo](https://github.com/mcdoerr/foxglove-mushr-extension) with the extension for details. 

**Description**
For the MuSHR package, we need 3 types of `PoseStamped` click-to-publish points. In our simulation, we need to be able to set the robot position, and for our real robot, we need to publish a point close to the robot. Finally, we need to be able to publish navigation goals for the robot on both simulation and real robot. The user can select the buttons on the MuSHR extension to change the way that the `PoseStamped` message is interpreted on the MuSHR backend for our 3 different pose types.