Add Sideways Teleop Extension (#3)
**Public-Facing Changes**
Add the first version of a small extension to the marketplace that takes the command of an Xbox controller and publish a Twist message. See [this repo](https://github.com/rscova/foxglove-sideways-teleop-extension) with the extension for details.

**Description**
Basically, we first detect if an Xbox controller is connected via USB. If it is connected, the extension starts to publish the Twist message with the value of the left and right sticks. Left for linear.x speed and right for angular.z speed.