CI: replace validation of zip compression with create-foxglove-extension minimum version (#37)
In #32, I added a CI check to validate that .foxe zip files actually use
compression (a bug fixed in
https://github.com/foxglove/create-foxglove-extension/pull/147).
However, it turns out this check was broken because when loading a zip
file with JSZip, `zipObj.options.compression` is not set even if the
file was compressed. This PR replaces it with an alternative approach
where we read the version of `create-foxglove-extension` from the
extension's package.json and validate that it's >=1.0.3.