Added WrenchStamped panel. (#45)
I have created the panel that visualize to WrenchStamped.msg.
https://docs.ros2.org/foxy/api/geometry_msgs/msg/WrenchStamped.html

It cannot be overlaid on the standard 3D panel, so I used a separate
panel, to display it.
(If there is a way to implement this, please let me know.)
Actually, I am thinking that an official feature should be provided that
allows Wrench data to be overlaid on the 3D panel.

I made it match the display in Rviz.

https://github.com/user-attachments/assets/4b5fdd5b-4b9b-4d0a-994c-dc39a42dc8b8